requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.0
aiosignal==1.3.2
aiounittest==1.3.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
azure-cognitiveservices-speech==1.44.0
azure-cognitiveservices-vision-computervision==0.9.1
azure-common==1.1.28
azure-core==1.33.0
azure-cosmos==4.7.0
azure-identity==1.23.0
azure-keyvault-secrets==4.9.0
azure-storage-blob==12.7.0
azure-storage-queue==12.4.0
Babel==2.9.1
botbuilder-azure==4.16.2
botbuilder-core==4.16.2
botbuilder-dialogs==4.16.2
botbuilder-schema==4.16.2
botframework-connector==4.16.2
botframework-streaming==4.16.2
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.2.1
constantly==23.10.4
contourpy==1.3.2
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
daphne==4.2.0
datedelta==1.4
Django==5.0.14
django-allauth==65.7.0
django-extensions==4.1
django-filter==25.1
django-phonenumber-field==8.1.0
django-tables2==2.7.5
django-widget-tweaks==1.5.0
django_injector==0.3.1
djangorestframework==3.16.0
dotenv==0.9.9
emoji==1.7.0
ffmpeg-python==0.2.0
fonttools==4.57.0
frozenlist==1.6.0
future==1.0.0
grapheme==0.6.0
graphviz==0.20.3
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
injector==0.22.0
isodate==0.7.2
jsonpickle==1.4.2
kaleido==0.2.1
kiwisolver==1.4.8
msal==1.32.0
msal-extensions==1.3.1
msgpack==1.1.1
msrest==0.7.1
mssql==1.0.1
mssql-django==1.5
multidict==6.4.4
multipledispatch==1.0.0
narwhals==1.36.0
numpy==2.2.5
oauthlib==3.2.2
packaging==25.0
phonenumbers==9.0.3
pillow==11.2.1
plotly==6.0.1
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydot==3.0.4
pydub==0.25.1
pydyf==0.11.0
PyJWT==2.10.1
pyodbc==5.2.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
recognizers-text==1.0.2a2
recognizers-text-choice==1.0.2a2
recognizers-text-date-time==1.0.2a2
recognizers-text-number==1.0.2a2
recognizers-text-number-with-unit==1.0.2a2
redis==6.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==25.5.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
weasyprint==65.1
webencodings==0.5.1
yarl==1.20.0
zope.interface==7.2
zopfli==0.2.3.post1
